comparison

+++ src/texture.rs
@@ -4,11 +4,11 @@ use crate::util::{color::Color, perlin::Perlin, Point3};
 use image::{DynamicImage, GenericImageView};
 
 pub trait Texture: Send + Sync {
-    fn sample(&self, u: f64, v: f64, point: Point3) -> Color;
+    fn sample(&self, u: f32, v: f32, point: Point3) -> Color;
 }
 
 impl Texture for Color {
-    fn sample(&self, _u: f64, _v: f64, _point: Point3) -> Color {
+    fn sample(&self, _u: f32, _v: f32, _point: Point3) -> Color {
         *self
     }
 }
@@ -16,13 +16,13 @@ impl Texture for Color {
 /// A Spatial Texture; Does not map to non-cartesian texture coordinate spaces.
 pub struct CheckerTexture {
     /// 1.0 / scale
-    pub inv_scale: f64,
+    pub inv_scale: f32,
     pub even: Arc<dyn Texture>,
     pub odd: Arc<dyn Texture>,
 }
 
 impl CheckerTexture {
-    pub fn new(scale: f64, even: impl Texture + 'static, odd: impl Texture + 'static) -> Self {
+    pub fn new(scale: f32, even: impl Texture + 'static, odd: impl Texture + 'static) -> Self {
         Self {
             inv_scale: 1.0 / scale,
             even: Arc::new(even),
@@ -32,7 +32,7 @@ impl CheckerTexture {
 }
 
 impl Texture for CheckerTexture {
-    fn sample(&self, u: f64, v: f64, point: Point3) -> Color {
+    fn sample(&self, u: f32, v: f32, point: Point3) -> Color {
         let x_int = (self.inv_scale * point.x).floor() as i32;
         let y_int = (self.inv_scale * point.y).floor() as i32;
         let z_int = (self.inv_scale * point.z).floor() as i32;
@@ -46,7 +46,7 @@ impl Texture for CheckerTexture {
 }
 
 impl Texture for DynamicImage {
-    fn sample(&self, u: f64, v: f64, _point: Point3) -> Color {
+    fn sample(&self, u: f32, v: f32, _point: Point3) -> Color {
         // DEBUGGING
         if self.height() == 0 {
             return Color::new(0.0, 1.0, 1.0);
@@ -55,32 +55,32 @@ impl Texture for DynamicImage {
         let u = u.clamp(0.0, 1.0);
         let v = 1.0 - v.clamp(0.0, 1.0);
 
-        let i = (u * self.width() as f64) as u32;
-        let j = (v * self.height() as f64) as u32;
+        let i = (u * self.width() as f32) as u32;
+        let j = (v * self.height() as f32) as u32;
 
         let color = self.get_pixel(i, j).0;
         Color::new(
-            color[0] as f64 / 256.0,
-            color[1] as f64 / 256.0,
-            color[2] as f64 / 256.0,
+            color[0] as f32 / 256.0,
+            color[1] as f32 / 256.0,
+            color[2] as f32 / 256.0,
         )
     }
 }
 
-pub struct NoiseTexture(Perlin, f64);
+pub struct NoiseTexture(Perlin, f32);
 
 impl NoiseTexture {
     pub fn new() -> Self {
         Self(Perlin::new(), 1.0)
     }
 
-    pub fn scaled(scale: f64) -> Self {
+    pub fn scaled(scale: f32) -> Self {
         Self(Perlin::new(), scale)
     }
 }
 
 impl Texture for NoiseTexture {
-    fn sample(&self, _u: f64, _v: f64, point: Point3) -> Color {
+    fn sample(&self, _u: f32, _v: f32, point: Point3) -> Color {
         let s = self.1 * point;
         Color::ONE * 0.5 * (1.0 + (s.z + s.x + s.y + 10.0 * self.0.turb(s, 7)).sin())
     }
